Changed package name
Also added tooltips
