+++ app/src/main/java/com/codechefvit/bookmarked/Favicon/Icon.java
@@ -1,0 +1,87 @@
+
+package com.codechefvit.bookmarked.Favicon;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+public class Icon {
+
+    @SerializedName("url")
+    @Expose
+    private String url;
+    @SerializedName("width")
+    @Expose
+    private Integer width;
+    @SerializedName("height")
+    @Expose
+    private Integer height;
+    @SerializedName("format")
+    @Expose
+    private String format;
+    @SerializedName("bytes")
+    @Expose
+    private Integer bytes;
+    @SerializedName("error")
+    @Expose
+    private Object error;
+    @SerializedName("sha1sum")
+    @Expose
+    private String sha1sum;
+
+    public String getUrl() {
+        return url;
+    }
+
+    public void setUrl(String url) {
+        this.url = url;
+    }
+
+    public Integer getWidth() {
+        return width;
+    }
+
+    public void setWidth(Integer width) {
+        this.width = width;
+    }
+
+    public Integer getHeight() {
+        return height;
+    }
+
+    public void setHeight(Integer height) {
+        this.height = height;
+    }
+
+    public String getFormat() {
+        return format;
+    }
+
+    public void setFormat(String format) {
+        this.format = format;
+    }
+
+    public Integer getBytes() {
+        return bytes;
+    }
+
+    public void setBytes(Integer bytes) {
+        this.bytes = bytes;
+    }
+
+    public Object getError() {
+        return error;
+    }
+
+    public void setError(Object error) {
+        this.error = error;
+    }
+
+    public String getSha1sum() {
+        return sha1sum;
+    }
+
+    public void setSha1sum(String sha1sum) {
+        this.sha1sum = sha1sum;
+    }
+
+}
