--- app/src/main/java/com/example/bookmark/Backend/Posts.java
@@ -1,56 +1,0 @@
-package com.example.bookmark.Backend;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Posts {
-
-    public Posts(String id){
-        this.id=id;
-    }
-
-    @SerializedName("id")
-    @Expose
-    private String id;
-    @SerializedName("url")
-    @Expose
-    private String url;
-    @SerializedName("created")
-    @Expose
-    private String created;
-    @SerializedName("user")
-    @Expose
-    private String user;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getUrl() {
-        return url;
-    }
-
-    public void setUrl(String url) {
-        this.url = url;
-    }
-
-    public String getCreated() {
-        return created;
-    }
-
-    public void setCreated(String created) {
-        this.created = created;
-    }
-
-    public String getUser() {
-        return user;
-    }
-
-    public void setUser(String user) {
-        this.user = user;
-    }
-}
