--- app/src/main/java/com/example/bookmark/Backend/Users.java
@@ -1,70 +1,0 @@
-package com.example.bookmark.Backend;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Users {
-
-    public Users(String email, String username, String password) {
-        this.email = email;
-        this.username = username;
-        this.password = password;
-    }
-
-
-    @SerializedName("email")
-    @Expose
-    private String email;
-    @SerializedName("username")
-    @Expose
-    private String username;
-    @SerializedName("password")
-    @Expose
-    private String password;
-    @SerializedName("auth_token")
-    @Expose
-    private String auth_token;
-    @SerializedName("id")
-    @Expose
-    private int id;
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getAuth_token() {
-        return auth_token;
-    }
-
-    public void setAuth_token(String auth_token) {
-        this.auth_token = auth_token;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-}
