+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -6,17 +6,17 @@
 
     <fragment
         android:id="@+id/navigation_home"
-        android:name="com.example.bookmark.CategoryFragments.home.HomeFragment"
+        android:name="com.codechefvit.bookmarked.CategoryFragments.home.HomeFragment"
         android:label="@string/title_home" />
 
     <fragment
         android:id="@+id/navigation_dashboard"
-        android:name="com.example.bookmark.CategoryFragments.dashboard.DashboardFragment"
+        android:name="com.codechefvit.bookmarked.CategoryFragments.dashboard.DashboardFragment"
         android:label="@string/title_dashboard"/>
 
     <fragment
         android:id="@+id/navigation_notifications"
-        android:name="com.example.bookmark.CategoryFragments.notifications.NotificationsFragment"
+        android:name="com.codechefvit.bookmarked.CategoryFragments.notifications.NotificationsFragment"
         android:label="@string/title_notifications" />
 
 </navigation>
